Put extensions in 1 file
For smaller projects this improves search/scannability and reduces the number of files in the project.

While Objective-C standards call for putting extensions in their own file with the naming convention "X+Y.filename", Swift extensions are better in the same file, or split into different files for different types (i.e. "StringExtensions.swift").
