--- Blockstack/Classes/String+Base64.swift
@@ -1,28 +1,0 @@
-//
-//  String+Base64.swift
-//  Blockstack
-//
-//  Created by Yukan Liao on 2018-04-19.
-//
-
-import Foundation
-import UIKit
-
-extension String {
-    
-    func fromBase64() -> String? {
-        guard let data = Data(base64Encoded: self, options: Data.Base64DecodingOptions(rawValue: 0)) else {
-            return nil
-        }
-        
-        return String(data: data as Data, encoding: String.Encoding.utf8)
-    }
-    
-    func toBase64() -> String? {
-        guard let data = self.data(using: String.Encoding.utf8) else {
-            return nil
-        }
-        
-        return data.base64EncodedString(options: Data.Base64EncodingOptions(rawValue: 0))
-    }
-}
