--- Blockstack/Classes/String+LeftPadding.swift
@@ -1,19 +1,0 @@
-//
-//  String+LeftPadding.swift
-//  Blockstack
-//
-//  Created by Yukan Liao on 2018-04-11.
-//
-
-import Foundation
-
-extension String {
-    func leftPadding(toLength: Int, withPad character: Character) -> String {
-        let stringLength = self.count
-        if stringLength < toLength {
-            return String(repeatElement(character, count: toLength - stringLength)) + self
-        } else {
-            return String(self.suffix(toLength))
-        }
-    }
-}
