--- Blockstack/Classes/URL+QueryParams.swift
@@ -1,22 +1,0 @@
-//
-//  URL+QueryParams.swift
-//  Blockstack
-//
-//  Created by Yukan Liao on 2018-04-11.
-//
-
-import Foundation
-
-extension URL {
-    
-    public var queryParameters: [String: String]? {
-        guard let components = URLComponents(url: self, resolvingAgainstBaseURL: true), let queryItems = components.queryItems else {
-            return nil
-        }
-        var parameters = [String: String]()
-        for item in queryItems {
-            parameters[item.name] = item.value
-        }
-        return parameters
-    }
-}
